Got the pop-up to work
Fixed the pop up after it wasn't working. Pushing this so that I don't lose this progress.

Flask not loading for me though, so I can't see the contents of the form properly. If anyone has a chance, can you please try the explore.html page and click the + icon button and see if it works?