RequestPage functionality
I've added RequestPage functionality, but I'm not sure how to test it, as my Flask is not working and I can't test it. Also been having some errors accessing the latest code, eg. when Jessica made a merge I could not see it, so I've had to manually copy the code a few times.

Have a few more things to fix, but almost done.

Can someone please help me test this?
- not sure how to add in functionality for answer question pop-up to redirect to login page if the user is not logged in.